[@mantine/hooks]: Extract common storage function for localStorage and sessionStorage.
