--- docs/src/docs/hooks/use-session-storage.mdx
@@ -1,107 +1,0 @@
----
-group: 'mantine-hooks'
-package: '@mantine/hooks'
-title: 'use-session-storage'
-category: 'state'
-order: 1
-slug: /hooks/use-session-storage/
-description: 'Use sessionStorage value as react state'
-import: "import { useSessionStorage } from '@mantine/hooks';"
-docs: 'hooks/use-session-storage.mdx'
-source: 'mantine-hooks/src/use-session-storage/use-session-storage.ts'
----
-
-## Usage
-
-use-session-storage allows you to use value from `sessionStorage` as react state.
-Hook works exactly the same as `useState`, but also writes the value to session storage:
-
-```tsx
-import { useSessionStorage } from '@mantine/hooks';
-
-// hook will read value from sessionStorage.getItem('color-scheme')
-// if sessionStorage is not available or value at given key does not exist
-// 'dark' will be assigned to value variable
-const [value, setValue] = useSessionStorage({ key: 'color-scheme', defaultValue: 'dark' });
-
-// Value is set both to state and sessionStorage at 'color-scheme'
-setValue('light');
-
-// You can also use callback like in useState hook to set value
-setValue((current) => (current === 'dark' ? 'light' : 'dark'));
-```
-
-## Serialize/deserialize json
-
-By default, hook will serialize/deserialize data with `JSON.stringify`/`JSON.parse`.
-If you need to store data in session storage that cannot be serialized with `JSON.stringify`
-– provide your own serialization handlers:
-
-```tsx
-const [value, setValue] = useSessionStorage({
-  key: 'color-scheme',
-  serialize: (value) => {
-    /* return value serialized to string */
-  },
-  deserialize: (SessionStorageValue) => {
-    /* parse sessionStorage string value and return value */
-  },
-});
-```
-
-## Usage with superjson
-
-[superjson](https://github.com/blitz-js/superjson) is compatible with `JSON.stringify`/`JSON.parse` but works for `Date`, `Map`, `Set` and `BigInt`:
-
-```tsx
-import superjson from 'superjson';
-
-const [value, setValue] = useSessionStorage({
-  key,
-  defaultValue,
-  serialize: superjson.stringify,
-  deserialize: (str) => (str === undefined ? defaultValue : superjson.parse(str)),
-});
-```
-
-## TypeScript
-
-### Definition
-
-```tsx
-interface UseSessionStorage<T> {
-  /** Session storage key */
-  key: string;
-
-  /** Default value that will be set if value is not found in session storage */
-  defaultValue?: T;
-
-  /** If set to true, value will be update is useEffect after mount */
-  getInitialValueInEffect?: boolean;
-
-  /** Function to serialize value into string to be save in session storage */
-  serialize?(value: T): string;
-
-  /** Function to deserialize string value from session storage to value */
-  deserialize?(value: string): T;
-}
-
-function useSessionStorage<T = string>({
-  key,
-  defaultValue,
-  deserialize,
-  serialize,
-  getInitialValueInEffect,
-}: UseSessionStorage<T>): readonly [T, (val: T | ((prevState: T) => T)) => void];
-```
-
-### Set value type
-
-You can specify value type same as in `useState` hook:
-
-```tsx
-const [value, setValue] = useSessionStorage<'dark' | 'light'>({
-  key: 'color-scheme',
-  defaultValue: 'light',
-});
-```
